Created Keyboard.h, removed references to `David J`
