+++ src/SimpleGui/Components/InputComponent.cpp
@@ -1,5 +1,5 @@
 //
-// Created by David J on 9/18/25.
+// Created by rondotdll on 9/18/25.
 //
 
 #include "InputComponent.h"
