+++ src/SimpleGui/Components/InputComponent.h
@@ -3,7 +3,7 @@
 #include "Component.h"
 
 //
-// Created by David J on 9/18/25.
+// Created by rondotdll on 9/18/25.
 //
