+++ src/SimpleGui/Components/TextBox.h
@@ -1,6 +1,6 @@
 #pragma once
 //
-// Created by David J on 9/4/25.
+// Created by rondotdll on 9/4/25.
 //
 
 #include "InputComponent.h"
