+++ src/SimpleGui/Components/Window.cpp
@@ -1,5 +1,5 @@
 //
-// Created by david on 8/24/2024.
+// Created by rondotdll on 8/24/2024.
 //
 
 #include "Window.h"
