+++ src/SimpleGui/helper.h
@@ -7,18 +7,6 @@
 #include "Types/UIGeometry.h"
 
 
-namespace Keyboard {
-
-#define BACKSPACE 0x08
-#define ENTER 0x0D
-#define SPACE 0x20
-#define UPPER_DOLLARSIGN 0x04
-
-  char readKey();
-
-}
-
-
 namespace SGui {
 // Find the true edges of the screen (useful for drawing around a screen border)
 // returns (width, height)
